MB-52130 Add build stubs for projector main
With regulator library being included in projector's code,
and regulator library being EE only feature, CE builds
would fail. Hence, two versions of main.go files are to
be maintained - One for enterprise and one for community.

The build stub at the top of the file takes care of this

Change-Id: I00ea5b0e4aedf62789eafe1c3c122c11cb5b260c