MB-52922 Use OSO for magma only for non-default scope/collection
Due to the performance regression found in MB-52857, GSI will only use
OSO for magma buckets for non-default scope/collection. This is only a
workaround till the underlying performance improvement for magma is
pending. The idea here is that in collection based data model, there is
a good chance that a single collection will not hold more than 50% of
bucket data and will benefit from OSO mode(magma OSO is about 50%
slower).

Change-Id: I6ac360a73f34ab2ccf3a34993f6537ec71b388ff