MB-53111 Use all disk snapshots on DCP rollback
It is better to try with all available disk snapshots on DCP
rollback. There can be cases where vbucket replica is behind indexer.
When such a replica gets promoted to active, it may ask to rollback
as indexer is asking to start using snapshot/seqno/vbuuid from its in-memory
state which the vbucket replica(now active) is not aware of.

This behavior is already in place if DCP asks indexer to rollback to 0.
With this change, even if DCP asks to rollback to a seqnum other than 0,
indexer will try to establish DCP streams with all available disk
snapshots rather than comparing the seqno and discarding more recent
snapshots.

Change-Id: I605bbc63437e4f93eb75760c1b530e3a52b2aa9a