MB-53183 Init snapAligned flag to false for timestamp
BP to 7.1.2

Explicitly set the snapAligned flag to false as the initial state.
HWT can be set from an earlier restartTs which is snap aligned
and the flag needs to be reset for new timestamp.

Also, set snapAligned to false if the check for snap aligned fails.

Change-Id: I5a6d508c4dd90795e120cc0366a2e81dcf66074d
(cherry picked from commit 762392034152c556da10f9f07f331f6879cd9847)