MB-53221 Use copies of snapshot timestamp when computing restartTs
BP to 6.6.x

During rollback, when indexer decides to rollback to a snapshot,
it reads the snapshot timestamp, updates lastRollbackTs in the
slice. The lastRollbackTs and restartTs point to same location
in memory.

Later restartTs is updated after computing new vbuuid's, the
lastRollbackTs is also updated with new vbuuid's. This breaks
the immutability constraint on lastRollbackTs.

This patch addresses the issue by making a copy of the snapshot
timestamps when updating restartTs and lastRollbackTs. This
decouples lastRollbackTs, restartTs. So, any updates to restartTs
will not be reflected in lastRollbackTs

Change-Id: I589d278572fda27844acdd382fa0f811dad96759
(cherry picked from commit 6400782d43a450f241f263f0087a202540aab360)