MB-52930 Recover index from disk on destination node during shard rebalance
a. After restore phase is completed, destination node will change the
   state of transfer token to ShardTokenRecoverShard
b. In this state, destination node will post REST requests to indexer
   to create and recover index instances
c. This patch adds the necessary infra to recover an index using a
   REST endpoint. The "/recoverIndexRebalance" endpoint takes care
   of this
d. Once the index is recovered, the state of the index is changed to
   INDEX_STATE_RECOVERED

Change-Id: Ibc0900ffbd7957a9ca4c704ccf63c66cdde07626