MB-52930 Handle ShardTokenReady state
a. Once all indexes are built on destination node, the state of the
   transfer token is moved to ShardStateReady
b. In this state, indexer will check the state of the sibling transfer
   token (A sibling transfer token is the one moving indexes from other
   node in the same subcluster)
c. The index instances on source subcluster will not be dropped until
   the sibling transfer token moves to ShardTokenReady state
d. Once sibling tranfser token moves to ShardTokenReady, source node
   will queue the index instances for dropping
e. After successfully dropping the index instances, the state of the
   transfer token will be changed to ShardTokenCommit
f. Master will change the ShardTokenCommit to ShardTokenDeleted and will
   delete the MetaKV entry of the transfer token

Change-Id: I8224a4fae20b0b0aea9d3405e2938a00d3db2988