MB-52927 Add SiblingTokenId state to transfer token
For shard rebalance, indexes move from one sub-cluster to another
As one transfer token is for one node to another, atleast 2 tokens
are required to move both master and replica indexes. The index
instances in source subcluster can be dropped only when both
master and replica indexes move to destination sub-cluster.  This
variable captures the tokenID of the sibling (For master token,
this will contain token ID of the replica and vice-versa). This
information is used to read the state of the sibling token to take
an action of dropping the indexes in source subcluster

Change-Id: Ie9dac095e50a73d388c910db17d578056e2e7b4b