MB-52928 Generate shard transfer tokens
Shard transfer tokens are generated using shardId's of
each index and the bucket, source and destination nodes
to which the indexes belong to.

Primary index is handled as a special case. Primary index
is present only on one shard. After generating tokens for
all secondary indexes, primary index is added to one of
these tokens that already contains the shardId of this
bucket and moving from same source to destination node.
If none of the tokens exist (i.e. a special case where the
shard contains only primary indexes), a new transfer token
is created

Change-Id: Icd0676896fef2737d5137ca1de750e960fb3b339