MB-53480 Use streamId to check for catchup pending
Current code uses index.Stream to check for catchup
pending. In some cases, the caller of this method
is passing "bucket" name as keyspaceId. If index is
in INIT_STREAM, then the check

index.Defn.KeyspaceId(streamId) == keyspaceId

would fail and catchupPending would be returned as false.

This will lead to indexer not starting inactive MAINT_STREAM.

Change-Id: Ibe2ee6cacadcdd575db2c142348fc0e15f6c857b