MB-52930 Make rebalance master decide on dropping of instances on source
The current rebalance logic works on the assumption that once
both transfer tokens of a tenant move to ShardTokenReady state
they can be dropped on source nodes.

However, due to the eventual consistency nature of metaKV, if
one indexer node in the subcluster receives a ShardTokenReady
state and indexer on other node crashes before receiving this
notification, then cluster affinity in the cluster will be
violated.

E.g., subcluster_1:indexer_1 saw that both transfer tokens have
moved to ReadyState and it drops that index instances on this node

subcluster_1:indexer_2 crashed before receiving the notification.
So, it sees that one token is in ReadyState while the other token
is in Recover state (or any other state). Therefore for this token,
the cleanup is expected to happen on destinaton node as transfer is
not complete during rebalance.

In this scenario, the sub-cluster affinity is violated as the indexes
remain on one node in source subcluster and other node in destination
subcluster.

To avoid this scenario, we serialise this operation at rebalance
master. Rebalance master will post a transfer token with a new state:
ShardTokenDropOnSource and populates the tokenIds for which instances
should be dropped on source. Source nodes on seeing this, will trigger
a drop.

If master were to crash after posting this drop, the token will still
be available in metaKV. So, cleanup will see this token and cleanup
instances on source. If this token is not available on metaKV, then
cleanup will drop indexes on destination nodes

Change-Id: I60aa750ea169e56d17734c678fc3326ee271e7c1