MB-54136 Cap the total number token in process at transferBatchSize
Prior to this fix, the initiateShardTransferAsMaster method did
not have an idea of how many tokens are currently in progress.
Due to this, if any transfer token has reached ShardTokenDeleted
state, tokens of size "transferBatchSize" will be published (as
the count always started from "0").

This patch fixes the issue by maintaining a global count

Change-Id: Icae81e6462022822f50e01a212493acb1f619b86