MB-54118 Distinguish between deferred and non-deferred indexes for shard rebalancer
Unlike normal rebalance algorithm, shard rebalancer must know
if an instance is deferred or not as it has to issue build
requests for all non-deferred index instances. Also, deferred
index instances will be recovered from disk (and not created
afresh) as they are already present in the shard on source
node.

Indexer can not rely on defn.Deferred flag alone as the flag
remains true even after index is built. To solve this problem,
the shard token generator populates the "InstStateAtRebal"
variable in the index definition.

When recovering index instances, indexer will recover an
instance as deferred instance if the state of index is
CREATED or READY and the "Deferred" flag in definition is
set to true. After recovery, the state of the index instance
will be moved to READY. In all other cases, the state of the
index instance will be moved to RECOVERED as the indexes will
be built during rebalance

Change-Id: I9501fdc5802160afd52c3c94473b1150015c39cc