MB-54137 Handle race between builder & shard rebalancer
Before this patch, when shard rebalancer sees an instance
in state CREATED or READY and defn.Deferred is set to true,
it would move the index state to READY after recovery.

If a build token is posted for this instance, then the
build token can start building the instance on destination
node after the state of the instances is moved to READY (as
destination nodes reads the tokens from metaKV). This build
command would fail as DDL is not supported in rebalance but
builder will continue to build the index.

Also, if a build token is posted, rebalancer is expected to
build the index as a part of rebalance. However, the index
with a build command should not be built along with all other
instances as this index will start from zero seqno.

To handle this case, when generating the shard transfer tokens,
planner will set the state of index instance to INITIAL if
a build token is posted for an index. During recovery, all
such instances will be moved to RECOVERED state (instead of
READY). This will make sure that builder will not pick the token
for processing (as builder only takes the READY tokens).

All the instances with INITIAL state are built in a separate
batch so that they can start with zero seqno. and not hinder
the recovery of other instances with non-zero seqno.

Change-Id: Ia91f93ebc732d851ce4a1e9778be2a9cf74b7722