MB-54183 Do not initiate Cancel for an already done rebalance
For a rebalance that is done, m.rebalancer will be set to nil.
If CancelTask is called while rebalance cleanup is in progress,
CancelTask will wait for rebalanceDone to finish & then initiates
Cancel(). But, rebalancer would be set to nil by this time
by the onRebalanceDoneLocked method. This would lead to a panic.

This patch fixes the issue by introducing a nil check before
calling Cancel() method.

Change-Id: I033fccf204fe6fa5e039874fe9ce19c48bb7109d