MB-53754 Add sibling index for replica repair cases
As replica repair in shard rebalance tries to copy the index
data from sibling node, add a new filed indicating the sibling
index. Executor would use this to generate transfer tokens with
sourceId as nodeUUID of the sibling index.

Also, for partitioned indexes are grouped so that all partitions
of an instance under repair share the same instanceId.
Earlier, planner was generating one instanceId for every partition

Change-Id: Ib6b71376dce1944000f22c4418a28c6dd28dee9f