MB-54116 Add serverless test - TestShardIdMapping
This test creates secondary, partitioned and primary indexes
on _default, c1 & c2 collections of bucket_1 and bucket_2.

Both deferred and non-deferred version of the indexes
are created and the shardID's of these indexes are
retrieved from getLocalIndexMetadata endpoint.

For each bucket, the number of shards created are verified.
For now, only 2 shards per bucket per node are supported

Change-Id: I2baa61a6f95fe3b144bb647668fcb6b67dd356e0