MB-53890 Add slice status monitor at shard transfer manager
Before each slice is closed, slice status monitor is notified
(via storage manager and shard_tranfer_manager) that the slice
is about to be closed.

When a DestroyShardId message comes to shard_transfer_manager,
it will wait until all slices of the shard are pending closure.
The shardId of a slice is retrieved from slice book-keeping.
The status of a slice is checked every 1 second. If the slice
is considered closed, it will be removed from book-keeping.

If there are no slices of a shard that are in the book-keeping,
it means all slices are closed. Hence, destroy can proceed safely.

Note: Indexer guarantees that before closing a slice, it will
notify sliceStatusMonitor. So, the ordering between slice
closure and destroy shard is guarateed i.e. destroy shard
arrives only after all indexes of a shard are dropped (using
/dropIndex command - which will invoke slice.Close() for
all slices)

Change-Id: I459983a29f10f7c10b8c4588102d9bc37de6060c