MB-53890 Populate shardIds at the time of slice initilization
As shardIds of a slice do not change after initilization,
pre-compute them at the time of shard initilization. This will
help to get the shardIds in cases where the stores are closed
but indexer needs to know the shardIds of a slice. Also, when
GetShardIds() is called, return from the slice book-keeping

Change-Id: Ic33a43504f9b347529b661ea4f5e5c5704face65