MB-54116 Add Infra for crash testing of shard rebalance
A new package testcode has been added which get's compiled
only for CI tests (The tag 2ici_test will be enabled only
for CI machine). For production, this method will not be
enabled and therefore becomes a no-op.

The purpose of these tests is to study shard rebalance
behaviour at various crash points

Change-Id: I37e3dcbd6d2b2c306227e740af35294fdfffec45