MB-54474 Use NewParsedValueWithOptions insted of json unmarshall for xattr processing
When using json.Unmarshall, projector will unmarshall all
fields of the xattr irrespective of it being used in index
or not. This is both CPU and Memory intensive.

This patch avoids that by using NewParsedValueWithOptions
where in query will parse the xattr fields only until required
values are found. It does not parse the entire byte slice.

Change-Id: Iaf60482ab7b6c573de60b66aea72f437c4adee29