MB-53892 use go module and fix monster tool path
we need prods and bags data from monster source dir.
when enabling go modules, this path is not directly available

this can be found in go path or go root. sample:
godeps/pkg/mod/github.com/prataprc/
monster@v0.0.0-20210210112206-07525cc27b6d

because of versioning here, manual find is hard
using go's build package to find it's directory

FYI: there is a catch here. if in future we require to move to
go's "vendor" based module resolution, that won't work as go
will not download the non-go files (aka bags and prods dir).
in that case we might want to consider moving the tool into
the indexing repo itself

Change-Id: I3a7711c3bc37e096b317fbb5c5319dc7abfeb55a