MB-54014 Include transfer progress in computation of rebalance progress
Rebalancer queries the getIndexStatus endpoint periodically to
retrieve the build progress of each index. The rebalance transfer
progress on each node is propagated as a response to getIndexStatus
call.

Shard Rebalancer will read the progress on each node depending on
the state of the transfer token and computes the build progress as
a weighted average. The transfer & restore phases have a weight of
35% each and recovery phase has been given 30% weight

Change-Id: If0de03a79cfc8e5264d4f38fb6bc20b777ea8ec4