MB-54634 Do not process collection drop messages twice
If a collection drop message is received while flush is in
progress, indexer will mark the instances as deleted & add
the instance list to streamKeyspaceIdPendCollectionDrop.

If a collection drop message is received again (as collection
drop message happens per vbucket), the local variables "scope"
and "collection" take default empty values. When the method
"deleteIndexInstOnDeletedKeyspace" is called with empty
collection value, it will delete all index instances on the
bucket which is undesirable.

This patch fixes the issue by comparing ths collection value
with empty string. If the collection value is empty, then it
means no instances belonging to the collectionId have been
found. In that case, indexer would return from the method

Change-Id: I0dcaf296d0e1d94a02a598277e257cd987a28022