MB-53977 Implement shard lock and unlock APIs
The shard lock & unlock API's are used to manage the meta
operations on a shard. E.g., when a shard is locked, no
new index (other than the rebalance issued indexes - rebalance
can create new indexes on the shard) can be created on the
shard until the shard is unlocked.

The semantics of shard locking are as follows:

a. Lock before initiating a shard transfer -> This will
   ensure that no new user issued index instances will be
   mapped on to the shard whose transfer is in progress
b. Lock the shard before starting to recover index instances
   from the shard -> This will ensure that only index
   instances that are a part of the shard will be recovered
   and no new user issued index instances can be added to the
   shard while recovery is in progress

Unlocking of the locked shard happens in the following cases:

a. Failed transfer -> On a failed transfer, shard rebalancer
   will unlock the shards and initiate a cleanup of the
   transferred data
b. Failed rebalance -> Rebalance can fail at multiple stages i.e.
   during restore, recovery etc. In such a case, the shards will
   be unlocked on source node if the source node is expected
   to hold the indexes. Otherwise, shards will be unlocked
   on the destination node
c. Successful transfer of bucket (From source sub-cluster to
   target sub-cluster). Once the token moves to ShardTokenCommit
   phase, destination node will unlock the shard for further DDL
   processing (As reaching ShardTokenCommit means that index
   instances are committed onto destination nodes)

Change-Id: I38c8e2c52ce26e0eda61a12b4c86b85ef82d34d8