MB-54408: Implement time sorting on Histogram (Log latency) distribution stats
Padded zeroes on the histogram interval value to open the possibility of sorting on string key, when the GetValue() of the Histogram stat is called. Iterates through the map in sorted order when StatsMap's byteSlice is being created and Marshalling automatically sorts the map while writing to file or when stats are being deduped.

Change-Id: I34fa6211c0791b0b9e1efc226b1035ceb4b6c630