MB-52931 Allow deferred index creation during rebalance
A new config, index.serverless.allowDDLDuringRebalance is added.
When indexer gets a create index request with type
CLUST_MGR_CREATE_INDEX_DDL, and if this config is set to true,
then indexer will allow DDL instead of rejecting it.

Lifecycle manager takes the decision on whether to allow a
DDL or not. If the DDL is on a bucket that is undergoing rebalance,
then the DDL is rejected and queued for back-ground execution.

If the DDL happens on a bucket that is not-undergoing rebalance
(or) on a bucket whose rebalance is finished, then the DDL is
allowed by lifecycle manager.

Change-Id: I27306dcfc672cabee1d1f11ca8235a83e14c1c90