MB-52931 Schedule DDL in back-ground on commit failure due to rebalance
A commit request to indexer can fail during rebalance when
rebalance is initiated after prepare is accepted. The prepareLock
on lifecycle manager will be reset when rebalance is initiated.

Lifecycle manager also checks the value of "RebalanceRunning"
in repo and rejects the commit request if rebalance is running.
In either case, metadata_provider will take the following
actions:

a. If atleast one node indexer node accepts the commit request,
then it means that a createCommandToken is posted and rebalancer
will handle the index creation/movement as a part of rebalance.
In this case, a schedule create token is not posted.

b. If none of the indexer nodes accept commit request (i.e. all
indexer nodes error out), then if the failure is due to rebalance,
then index posts a schedule create token for the index - If the
schedule create token does not already exist (i.e. in
asyncCreate == false cases)

Before posting a schedule create token, metadata provider will
cancel the prepareLock acquired at lifecycle manager in earlier
phases.

Change-Id: I4b1d173e595b2df13763fd2b182b88316852e215