MB-52931 Allow index build during rebalance
During a rebalance, if a build index command is received,
then indexer will attempt to build the index on source
node if:

a. Bucket is not a part of rebalance - Then the index
will be built immediately

b. Bucket is undergoing rebalance - Then the index will
be built once the rebalance of the bucket is finished.
This is because indexer will wait for the destination
node to commit to the ownership of the index before
attempting to build the index.

As build token is posted, builder will take care of building
the index in the back-ground.

Change-Id: If5d37c66e3725da3f084e8e210fa875ca2dbfbfc