MB-52931 Skip index creation or build on destination node for dropped index
During rebalance, if a drop index command is received, then
indexer would proceed to drop the index on destination node
as mentioned below:

a. Bucket is not a part of rebalance - In this case, the destination
node will not receive the drop index command. This will be a no-op
for destination node

b. Bucket is a part of rebalance - In this case, for rebalance
initiated requests, lifecycle manager would check the presence
of a deleteCommandToken in metaKV before attempting to create
index instance or build index instance. If a deleteCommandToken
exists with the defnId of the instance being created/build,
lifecycle manager would skip the index instance.

Since metaKV is a I/O call to ns_server - Lifecycle manager
enforces a timeout of 5 seconds. If the token can not be
read within 5 seconds, lifecycle manager considers that the
index exists and continues to create or build the index.
In such a case, lifecycleMgr::janitor takes the job of deleting
the index (Similar to the case where index gets dropped while
it is building on destination node)

Note-1: In serverless deployments, user is not exposed to replicas.
So, there can not be drop replica. Hence, only DropIndex at client
is possible. Client will always post of deleteCommandToken in metaKV
with the definition ID of the index before initiating drop)

Note-2: Currently, rebalance would fail if index is dropped on
destination node during rebalance. The rebalance logic needs to be
updated to handle these cases.

Change-Id: If98c9c9a463c8454e42ea78acf6b587458083c8c