MB-52931 Handle deletion of scope/collection/index during rebalance
If scope or collection or index are dropped while rebalance is
in progress, then lifecycle manager would drop the indexes.
In such case, rebalancer will check the response of the REST
calls to indexer and if it sees errors related to scope/
collection/index being dropped - then it will skip those
instances from further processing.

A new error "ErrIndexDeletedDuringRebal" is added to handle
errors when an index is deleted while rebalance is in progress.

Change-Id: Ic4b8e00ca1fee4a462d70e39fc81a536349708f7