MB-52931 Skip deleted indexes from waitForIndexState processing
When /recoverIndexRequest is sucessful, it means that an index
instance is updated in metadata with state INDEX_STATE_CREATED.
Later, the index state can be update to READY or RECOVERED or
INITIAL or CATCHUP or ACTIVE.

However, if an index is deleted, then the index will be cleaned
up from topology. Then the state of the index would either be
NIL or DELETED. In such cases, skip the instance in
waitForIndexState.

Change-Id: I189683483591ae379b7efc3c0c5ebd14997ba7b5