MB-53902 Add blob_storage_region to indexer settings
The bucket region information is mandatory for operating on
blob_storage_bucket (atleast for S3)

As control plane is the one creating the bucket (S3 or equivalent)
and passing on the bucket information to indexer via ns_server, it
is desirable to have control plane pass on the bucket region
information. Otherwise, indexer has to query the EC2 service to
query the region - which brings up unnecessary dependency for
indexer on EC2 service.

Change-Id: Icb4d49b8568928b74c23039abd5fdc680cd69a4c