MB-52931 Notify lifecycle manager on async recovery done
If index is dropped during async recovery (either due to
scope/collection drop (or) user issued a drop), then
indexer should defer the index deletion until recovery
is done - as plasma expects indexer to serialize slice
recovery and slice destruction.

To achieve this, lifecycle manager will mark the index
instances that are in async recovery. If any of these
instance are dropped while recovery is in progress, then
they will be dropped after recovery is done. Indexer
will notify that async recovery is done for an instance
after recovery is complete.

Also, metadata updates to the index after recovery are
skipped if index is dropped during async recovery

Change-Id: I11ced811c0519103ce8081b9e7c850c95e5633c0