MB-52931 Do not limit shard rebalance initiated builds
With support for DDL during rebalance, there can be
multiple initial index builds going on in the indexer.
This can lead to max_parallel_collection_builds
exhaustion and rebalancer might not be able to build
the index immediately.

As shard rebalance is expected to be fast, removing
the max_parallel_collection_builds limitation for
shard rebalance. The rationale is that a max of 2
collection strems are allowed during rebalance per
node (as serverless rebalance transfer batch size
is 4, only 2 buckets can be moved at a time based on
the tranfser token grouping done by shard rebalancer).
They should not add more load to the system as these
streams will only be catching up mutations.

Change-Id: I1fec5b4a46b301fa1c8df9abdfa5a60b2399a3b4