MB-53977 UseShardId as specified by rebalancer
During recovery, if plasma can not find a recovery point
for an index instance, it will reset the stores. When the
stores are reset, the plasma instance should be mapped
to the same shard on which it was transferred. Otherwise,
this can violate symmetric distribution of index placement
on the shards.

To enable this, indexer specifies the shardIds that are to
be used when initialising plasma instance so that reset
stores will still map the plasma instance to the shard on
which the instance is transferred.

Plasma's UseShardId config is used to address this.

Change-Id: I93def353b9d0ba35158739d12263c9ad5e68e9dc