MB-54915 Destroy index snapshot if index is deleted
The following sequence of events can lead to a leaked
snapshot:

a. User issues a drop index command
b. Flush is in progress for this instance
c. Because of flush in progress, indexer will mark the
   instance as DELETED, sends the new index instance map
   to all workers - including storage manager
d. When storage manager receives this message, it will
   clear the index from indexSnapMap
e. Storage manager calls updateSnapMapAndNotify after
   flush is done
f. If (e) happens after (d), then snapC would be nil
   as the index is deleted - Therefore, updateSnapMapAndNotify
   would return with out updating the snapshot container - But,
   it does not destroy the already created snapshot. Hence,
   there will always be one open snapshot

Change-Id: I0a3943c56b82a024ad75c1f31ee2e66ffb3a924a