MB-52931 Always allow rebalance initated drop requests
In many places, the reqCtx for DeleteIndex is not populated.
This makes it difficult to distinguish between a user issued
drop or rebalance issued drop.

With the support for DDL and rebalance, both types of requests
are possible at same time. While user issued requests needs
to be blocked in some cases (in cases like index deletion
happening when transfer is in progress), rebalance issued
requests are to be allowed in all cases (as rebalance is
orchestrating the transfer). Hence, populating reqCtx for
all DeleteIndex calls.

Change-Id: Idd55abeb594da1d0f1bd7cedc3b4eca4e34a3976