MB-52931 Drop index (not just instance) after async recovery is done
If an index drop is received while async recovery is in
progress, indexer will process the drop only after async
recovery is done. At this time, delete the index as there
is no concept of replica drop in serverless mode.

Otherwise, the index defnID will remain in meta and this
can lead to index creation failures when another index with
same name on same keyspace is created.

Change-Id: I743ebe269766fb8e13835a3fa5fbcc61bc74ef34