MB-55062 Skip destoying shards on dropIndex failure
If index is in async recovery, dropIndex would fail with
"ErrIndexInAsyncRecovery". In such a case, /dropIndex would
fail. Do not destroy shards when dropIndex fails as indexer
could still be in async recovery.

A long term solution would be to cancel recovery when dropIndex
happens. This requires changes in plasma code as well

Change-Id: Ib7637c6194907b3a328d02d07ef740e38c285a4f