MB-55077 Clear rebalance running flag at the end of rebalance
For instances created during rebalance, if resetStores happen,
then the plasma stores should be mapped to the same shards. To
facilitate this, indexer specifies UseShardId in mainstore and
backstore config.

The contract with plasma is to populate UseShardId flag only
during rebalance. If ResetStores() were to happen when rebalance
is not running, then indexer should not populate UseShardId in the
config.

To honour the contract, indexer keeps a track of rebalance with a
flag. This flag will be cleared either when shards are unlocked or
at the end of rebalance.

Change-Id: I55d62acc23f8dce764ca6713792054530224637e