MB-55114 Remove index metadata for missing instances
In general, all DDL operations via client go through watchers.
However, rebalance initiated DDL operations does not go
through watchers. There exists one race conditions during
rebalance cleanup where index scans can fail.

a. After updating the RState of a rebalance instance, indexer
on destination node crashes (before updating the transfer token
state)
b. Because the RState is updated, client will update the book
keeping and will consider the index available for scans
c. Due to indexer restart, watcher will terminate and attempt
to reconnect
d. While watcher is trying to sync with peer, destination node
has cleaned up the index (as transfesr token state is not yet
updated)
e. There are two actions on topology at this poing:
   (1) Updating the state of index to DELETED
   (2) Deleting the index from topology
f. When watcher is not active, it is possible that only the last
update to topology is received by watcher. This means, that the
state updation of DELETED is not propagated to watcher.
g. According to current client logic, the instance metadata is
cleaned up only if watcher receives DELETED notification. If
watcher misses this, then client will continue to consider the
index as active and route scans to the instance (which is deleted)

To avoid this race condition, client will keep a track of all index
definitions that are removed from the topology. For such instances,
it will re-update the instance metadata - so that metadata is always
updated with active instances

Change-Id: Ia09de16359c6d4930675fd0d8a876cb82dad421c