MB-55092 add listener for supvervisor command
bg: in serverless mode, pause resume manager should listen to certain
events - config update, indexer ready and shutdown. this is to keep config
in sync with the service and run crash recovery on restart.

exp: add a command message channel for pause resume. add a routine to
listen to commands from indexer and handlers for those commands

stubs: recoverFromCrash for pause resume manager

Change-Id: Ic9b176ea944475b41f8e1d816bbcf8cd6697ae03