MB-54907 Pass bucket region from indexer config to plasma
As an S3 bucket's region is a property associated with the
bucket, control plane is expected to pass on blob_storage_region
via ns_server config.

Shard rebalancer will read this config and persist the information
in transfer token. The same information will be passed on to
plasma during transfer and restore.

Change-Id: I364522a5f435713e79c6e15fefa4432b536b94c2