MB-55309 Set forceUnbalanced to false during cleanup
The purpose of calling onRebalanceDoneLOCKED() in cancel
prepare task is to cleanup indexer book-keeping and not
to change the state of the cluster. Hence set forceUnbalanced
to false.

Otherwise, rebalance button is enabled even after rebalance
(as indexer reports node is not balanced) and control plane
will trigger another rebalance.

Change-Id: I2b4b43669c3ee7683367eabe804f5deec505043c