MB-54860 Skip closing slices during rebalance transfer
slice.Close() and slice.Destroy() will change the metadata
of shard. While rebalance transfer is in progress, if a slice
is closed or destoryed, then rebalance can fail due to mismatch
in metadata.

To avoid this issue, indexer keeps a track of bucket transfer
phase. If an index is dropped while transfer is in progress,
then the slice closure and destruction is skipped. Once transfer
is complete, then the slice is closed and destroyed.

If rebalance fails while transfer is in progress and an index
is also dropped during transfer, then slice is closed during
rebalance cleanup.

Change-Id: I56a8eb76051d07a45021ef4ccdee308472585f70