MB-54860 Allow drop during rebalance transfer
When there is a user initiated dropIndex request during
rebalance transfer in progress, currently index metadata is
updated but indexer book-keeping is not deleted (i.e. slice
closure and destruction is skipped). This will lead to following
race condition:

a. Rebalance has started and is in PrepareTopologyChange
b. User initiated a dropIndex for an index instace
c. Lifecycle manager will update the state of the index
   instance to INDEX_STATE_DELETED but OnIndexDelete()
   will not be called due to rebalance in progress
d. During StartTopologyChange, planner will not consider
   the index instance as the state is updated to DELETED
e. After rebalance is done, rebalancer will destory the
   shard - But the slice of an index is not closed yet.
   This will lead to indexer panic - as shard is destroeyd
   while mutation processing is in progress

To avoid this race condition, dropIndex will initiate cleanup
at indexer as well. Indexer will update its book-keeping and
postpone the closure of slice. At the same time, the sliceStatus
monitor keeps a track of the slice closure. When destroy shard
happens, it will wait for the closure of slice before destroying
the shard.

Change-Id: I0a7758ed5bcf93cde252e2eae1c6fc7a051bdbe5