MB-54756 Set rebalance running to slices on source node
When indexer had to resetStores() for a plasma slice on source
node (Due to rollbackToZero), then setting rebalance running
flag will enable plasma to map the index instances on a locked
shard (as the shards would be locked for transfer).

The flag will be cleared either during shard unlock or at the
end of rebalance cleanup.

Change-Id: I104147ee56116f7b6744aedcc7cc118a4e9113d0