MB-54756 Route shard transfer response via stoarge manager
This will help storage manager to update transfer book-keeping
which will be used later to determine the status of transfer.

For shards that are in transfer, if any slice belonging to the
shard receives rollbackToZero, then plasma can determine the
status of the transfer by looking at its in-memory map and take
a decision on the rebalance transfer.

Change-Id: I6fc039f1872ee91b46cc507ad9ccae3aa94e2f0c