MB-54756 Cancel rebalance transfer before rolling back to zero
When there is a rollbackToZero during rebalance transfer, indexer
will delete existing plasma instances and creates new intances.
As this is going to chance metadata of the shard, do this operation
will lead to inconsistent data (and metadata) being transferred.

To avoid such issues, if indexer has to rollback a slice to zero
during transfer, it will first abort the transfer. Then it waits
for plasma to complete the transfer and then attempts to rollback
the slice to zero. This will ensure that indexer will not transfer
inconsistent data to destination node.

Change-Id: I80e2b733a0c01e03dcbf9ee0a8b12fcae7af8405