MB-56559 [BP 7.1.4] Redistribute index flag works on partitioned indexes as well
[BP MB-55517 to 7.1.4]

The redistribute_index flag now affects partitioned indexes as well. When the flag is false, both Partitioned and Non-partitioned Index are not redistributed across Nodes on rebalance ( for Index Node removal & Swap Rebalance the behaviour is not affected by the flag). When flag is true, the behaviour remains same as before

Change-Id: I81b220fde7a54d5ff5779847b4f14b414f7bedfd