MB-55517 Redistribute index flag works on partitioned indexes as well
[BP to 7.2.0]

The redistribute_index flag now affects partitioned indexes as well. When the flag is false, both Partitioned and Non-partitioned Index are not redistributed across Nodes on rebalance ( for Index Node removal & Swap Rebalance the behaviour is not affected by the flag). When flag is true, the behaviour remains same as before

Change-Id: I81b220fde7a54d5ff5779847b4f14b414f7bedfd